make_room_configurable: room <email> is obtained during application startup renamed scala packages renamed main App object updated Readme
