+++ README.md
@@ -1,1 +1,9 @@
-# smart-room-iot
+# start motion sensor
+## udp server to report to motion events need to become configurable
+python motionsensor.py 
+
+# start server app
+room is identified by an email address. example: CONF_374@<domain>.com
+room, exchange user and password are obtained as input from console during startup
+sbt runMain smart.SmartRoom
+
